requirements.txt
langgraph==0.2.*
langchain==0.3.*
langchain-openai==0.3.*
fastapi==0.115.*
uvicorn[standard]==0.30.*

openai==1.52.*

chromadb==0.5.*
pypdf==5.0.*
tiktoken==0.7.*

pydantic==2.9.*

python-dotenv==1.0.*
